add: 게시글 삭제 확인창

+++ frontWEB/static/js/noticeBoardContent.js
@@ -179,29 +179,32 @@ async function postReplyComment(upperId, textareaId){
 
 //삭제 버튼 클릭시
 deleteBtn.onclick = async function(){
-  var myHeaders = new Headers();
-  var raw = JSON.stringify({
-    "noticeNum": pageNum
-  });
-
-  myHeaders.append("x-access-token", localStorage.getItem("accessJWT"));
-  myHeaders.append("Content-Type", "application/json");
-
-  var requestOptions = {
-    method: 'DELETE',
-    headers: myHeaders,
-    body: raw,
-    redirect: 'follow'
-  };
-
-  console.log(requestOptions)
-  const deleteRes = await deleteAPI(hostAddress, 'app/notice/delete', requestOptions);
-  if(deleteRes.isSuccess == true) location.href = '/notice';
-  else{
-    alert(deleteRes.message);
-    location.href = '/notice';
-  }
+  if (confirm('게시글을 삭제하시겠습니까?')) {
+    var myHeaders = new Headers();
+    var raw = JSON.stringify({
+      "noticeNum": pageNum
+    });
+  
+    myHeaders.append("x-access-token", localStorage.getItem("accessJWT"));
+    myHeaders.append("Content-Type", "application/json");
+  
+    var requestOptions = {
+      method: 'DELETE',
+      headers: myHeaders,
+      body: raw,
+      redirect: 'follow'
+    };
+  
+    const deleteRes = await deleteAPI(hostAddress, 'app/notice/delete', requestOptions);
+    if(deleteRes.isSuccess == true) location.href = '/notice';
+    else{
+      alert(deleteRes.message);
+      location.href = '/notice';
+    }
+  } 
 }
+
+
 //수정 버튼 클릭시
 modifyBtn.onclick = function(){
   alert("기능 준비중입니다.\n관리자 전준휘에게 문의하여 수정하실 수 있습니다.")
